event_ez/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
bandit = ">=1.8.3"
fastapi = {extras = ["all"], version = "^0.111.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4"}
pre-commit = ">=4.1.0"
pyjwt = ">=2.10.1"
python-dotenv = ">=1.0.1"
ruff = ">=0.9.7"
sqlalchemy = ">=2.0.38"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"